requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kubernetes==28.1.0
passlib==1.7.4
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pyyaml==6.0.1

